Revert "Use ImageDecoder for BitmapDrawable"
Bug: 72381918
Test: Covered by existing tests

This CL seems to have broken the theme tests. In addition, it changes
the behavior of a couple of cases of passing InputStreams to the
framework. Previously, the framework used BitmapFactory, which did not
close the InputStreams, but ImageDecoder does.

Planning to reland along with the fix for closing in ag/3497523 and the
fix for the theme tests.

This reverts commit 06f4d1c290c584f30059ce93934c6e73a903a6aa.

Change-Id: I4ac6d0f2e1e2bb0925ae71c141bfe8a0d37e6a16