Re-initialize the brightness slider when out of sync.
Whenever the brightness controller stops listening to brightness events,
it could get out of sync and so will animate when next opened and it
re-registers. We want to avoid this animation since it looks odd to the
user so we always re-initialize whenever we might get out of sync.

Bug: 72635553
Test: manual
Change-Id: I3a7e05e18652af33511a5de5a8ebeee01d2f0349
(cherry picked from commit 1cae8150daac3f1aaf632a68a4710bc7e49c57f2)