Move bugreport storage enabling into OS.
It's convoluted to enable/disable bugreport storage based on the
"power menu" shortcut, and it's frustrating for dogfooders to have
to manually cycle this setting once to enable the provider.

So move the enable/disable logic directly into the OS, toggle it
when the broader DEVELOPMENT_SETTINGS_ENABLED changes, and update
to match state at device boot.

Test: builds, boots
Bug: 32540478
Change-Id: Iadfa106142fa3e252dfe6adb9712f7c67b8e9f65