Fix missing JavaDoc of IMM#toggleSoftInput()
This CL addresses a typo in the JavaDoc of IMM#toggleSoftInput(),
which has prevented it from being treated as a valid JavaDoc.

Note that this CL removes @hide simply because it is too late.  The
method in question has already been exposed to developers since API
level 3 (Cupcake) because of the typo.

Bug: 13002424
Test: make -j doc-comment-check-docs
Change-Id: I3e42d433f20c56ee63373bb390aad59e3bd7ec4f