Revert "Plumb IME subtype change from IMMS to IMS."
This reverts commit 7a6748cd8fc8cdb6f5355ed3eca01637a9d990d1 [1].

Reason for revert:
This is the part 3 of a series of reverts to unlaunch Bug 25752812,
which aimed to improve UX but did not work well.

Note that since this was just a plumbing CL, reverting this CL itself
should introduce no user-visible behavior change.

See I7a2ed6dd39dcd8207d3d94e12cd01d5d67ba4bb5 for the detailed reason
of revert.

 [1]: I58e71ffce9ac9131551a00dd35e24235dadfef87

Bug: 66498367
Test: Manually done
Change-Id: I3e2b9a9faf5234e53d20953886d9d55af680087d