Enable optional divider in AccessPointPreference
Create a new preference layout based on layout_preference_two_target
that will allow for the optional use of a second target with a ripple
animation that stops at the hairline divider as well as a toggle for the
hairline divider.

This will enable ConnectedAccessPointPreference to
display a split between the left and right half when a Captive Portal
signin is pending.

Bug: 63929546
Bug: 68031656
Test: make RunSettingsLibRoboTests
Test: manual testing with Captive Portal and normal WiFi networks.

Change-Id: I5c8c513290ed04a8dcc38c3f9d6c149c9cca08aa
(cherry picked from commit 70476fc1505c9d20f2f430588b8b0b3e7ea973b1)